+++ src/main/java/com/portfolio/portfolio/repository/EstudioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EstudioRepository  extends JpaRepository<Estudio, Integer> {
+public interface EstudioRepository  extends JpaRepository<Estudio, Long> {
 }
